chore: major refactoring/restructuring, adding fake requests to getTestResult
