+++ src/xcode/ENA/ENA/Source/Client/HTTP Client/URLSession+Convenience.swift
@@ -22,14 +22,14 @@ extension URLSession {
 
 	// This method executes HTTP GET requests.
 	func GET(_ url: URL, completion: @escaping Completion) {
-		response(for: URLRequest(url: url), completion: completion)
+		response(for: URLRequest(url: url), isFake: false, completion: completion)
 	}
 
 	// This method executes HTTP POST requests.
 	func POST(_ url: URL, completion: @escaping Completion) {
 		var request = URLRequest(url: url)
 		request.httpMethod = "POST"
-		response(for: request, completion: completion)
+		response(for: request, isFake: false, completion: completion)
 	}
 
 	// This method executes HTTP POST with HTTP BODY requests.
@@ -38,7 +38,7 @@ extension URLSession {
 		request.httpMethod = "POST"
 		request.httpBody = body
 		request.setValue("application/json", forHTTPHeaderField: "Content-Type")
-		response(for: request, completion: completion)
+		response(for: request, isFake: false, completion: completion)
 	}
 
 	// This method executes HTTP requests.
@@ -47,9 +47,15 @@ extension URLSession {
 	// - if there is either no HTTP body and/or HTTPURLResponse it aborts
 	func response(
 		for request: URLRequest,
+		isFake: Bool,
 		completion: @escaping Completion
 	) {
 		dataTask(with: request) { data, response, error in
+			guard !isFake else {
+				completion(.failure(.fakeResponse))
+				return
+			}
+
 			if let error = error {
 				completion(.failure(.httpError(error)))
 				return
@@ -101,6 +107,7 @@ extension URLSession.Response {
 		case regTokenNotExist
 		case invalidResponse
 		case serverError(Int)
+		case fakeResponse
 	}
 
 	typealias Completion = (Result<URLSession.Response, Failure>) -> Void
